# backend/openapi.json
{
 "components": {
  "schemas": {
...
   "Event": {
    "properties": {
     "category": {
      "enum": [
       "kernel_log",
       "pci_error",
       "pci_link",
       "usb_error",
       "usb_transfer",
       "hotplug",
       "driver_state",
       "firmware",
       "association",
       "phy",
       "mlo",
       "reset",
       "power",
       "scan",
       "regulatory",
       "trace",
       "telemetry",
       "netdev",
       "diagnostic",
       "system"
      ],
      "title": "Category",
      "type": "string"
     },
     "data": {
      "additionalProperties": true,
      "title": "Data",
      "type": "object"
     },
     "demo": {
      "default": false,
      "title": "Demo",
      "type": "boolean"
     },
     "device_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Device Id"
     },
     "explanation": {
      "title": "Explanation",
      "type": "string"
     },
     "id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Id"
     },
     "kind": {
      "title": "Kind",
      "type": "string"
     },
     "raw": {
      "title": "Raw",
      "type": "string"
     },
     "session": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Session"
     },
     "severity": {
      "default": "info",
      "enum": [
       "debug",
       "info",
       "notice",
       "warning",
       "error",
       "critical"
      ],
      "title": "Severity",
      "type": "string"
     },
     "source": {
      "title": "Source",
      "type": "string"
     },
     "summary": {
      "title": "Summary",
      "type": "string"
     },
     "ts_boottime_ns": {
      "description": "Canonical timestamp: CLOCK_BOOTTIME nanoseconds",
      "title": "Ts Boottime Ns",
      "type": "integer"
     },
     "ts_raw": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "Original timestamp in its own clock domain",
      "title": "Ts Raw"
     },
     "ts_source": {
      "enum": [
       "tracefs_boot",
       "tracefs_mono",
       "kernel_printk",
       "journal_monotonic",
       "receive_boottime",
       "poll_boottime",
       "netlink_boottime",
       "usbmon_monotonic",
       "recorded"
      ],
      "title": "Ts Source",
      "type": "string"
     },
     "ts_wall": {
      "description": "Derived wall clock (unix seconds) - display only",
      "title": "Ts Wall",
      "type": "number"
     }
    },
    "required": [
     "id",
     "ts_boottime_ns",
     "ts_source",
     "ts_raw",
     "ts_wall",
     "source",
     "device_id",
     "severity",
     "category",
     "kind",
     "summary",
     "explanation",
     "raw",
     "data",
     "demo",
     "session"
    ],
    "title": "Event",
    "type": "object"
   },
   "EventPage": {
    "properties": {
     "events": {
      "items": {
       "$ref": "#/components/schemas/Event"
      },
      "title": "Events",
      "type": "array"
     },
     "next_after_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next After Id"
     },
     "server_boottime_ns": {
      "title": "Server Boottime Ns",
      "type": "integer"
     },
     "server_wall": {
      "title": "Server Wall",
      "type": "number"
     }
    },
    "required": [
     "events",
     "next_after_id",
     "server_boottime_ns",
     "server_wall"
    ],
    "title": "EventPage",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "TelemetrySample": {
    "description": "A periodic numeric sample (for charts). Not an event; kept in a bounded ring.",
    "properties": {
     "demo": {
      "default": false,
      "title": "Demo",
      "type": "boolean"
     },
     "device_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Device Id"
     },
     "series": {
      "title": "Series",
      "type": "string"
     },
     "ts_boottime_ns": {
      "title": "Ts Boottime Ns",
      "type": "integer"
     },
     "ts_wall": {
      "title": "Ts Wall",
      "type": "number"
     },
     "values": {
      "additionalProperties": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ]
      },
      "title": "Values",
      "type": "object"
     }
    },
    "required": [
     "ts_boottime_ns",
     "ts_wall",
     "device_id",
     "series",
     "values",
     "demo"
    ],
    "title": "TelemetrySample",
    "type": "object"
   },
...
   "TraceStart": {
    "properties": {
     "buffer_kb": {
      "default": 4096,
      "maximum": 16384.0,
      "minimum": 64.0,
      "title": "Buffer Kb",
      "type": "integer"
     },
     "events": {
      "description": "'group/event' or 'group/*'",
      "items": {
       "type": "string"
      },
      "maxItems": 256,
      "minItems": 1,
      "title": "Events",
      "type": "array"
     },
     "max_seconds": {
      "default": 600,
      "maximum": 86400.0,
      "minimum": 5.0,
      "title": "Max Seconds",
      "type": "integer"
     }
    },
    "required": [
     "events"
    ],
    "title": "TraceStart",
    "type": "object"
   },
...
   "UsbmonStart": {
    "properties": {
     "buses": {
      "anyOf": [
       {
        "items": {
         "type": "integer"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "description": "default: buses of present USB wireless devices",
      "title": "Buses"
     },
     "max_seconds": {
      "default": 300,
      "maximum": 3600.0,
      "minimum": 5.0,
      "title": "Max Seconds",
      "type": "integer"
     }
    },
    "title": "UsbmonStart",
    "type": "object"
   },
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Wireless Hardware Debugger",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/captures/{cid}/events": {
   "get": {
    "operationId": "capture_events_api_v1_captures__cid__events_get",
    "parameters": [
     {
      "in": "path",
      "name": "cid",
      "required": true,
      "schema": {
       "title": "Cid",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "offset",
      "required": false,
      "schema": {
       "default": 0,
       "title": "Offset",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 2000,
       "maximum": 20000,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/Event"
         },
         "title": "Response Capture Events Api V1 Captures  Cid  Events Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Capture Events"
   }
  },
...
  "/api/v1/events": {
   "get": {
    "operationId": "list_events_api_v1_events_get",
    "parameters": [
     {
      "in": "query",
      "name": "device_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Device Id"
      }
     },
     {
      "in": "query",
      "name": "categories",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Categories"
      }
     },
     {
      "in": "query",
      "name": "min_severity",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "debug",
          "info",
          "notice",
          "warning",
          "error",
          "critical"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Min Severity"
      }
     },
     {
      "in": "query",
      "name": "sources",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Sources"
      }
     },
     {
      "in": "query",
      "name": "kinds",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Kinds"
      }
     },
     {
      "in": "query",
      "name": "text",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Text"
      }
     },
     {
      "in": "query",
      "name": "since_ns",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Since Ns"
      }
     },
     {
      "in": "query",
      "name": "until_ns",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Until Ns"
      }
     },
     {
      "in": "query",
      "name": "after_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "After Id"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 500,
       "maximum": 10000,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "order",
      "required": false,
      "schema": {
       "default": "desc",
       "pattern": "^(asc|desc)$",
       "title": "Order",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventPage"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Events"
   }
  },
  "/api/v1/events/{event_id}": {
   "get": {
    "operationId": "get_event_api_v1_events__event_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "event_id",
      "required": true,
      "schema": {
       "title": "Event Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Event"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Event"
   }
  },
...
  "/api/v1/telemetry": {
   "get": {
    "operationId": "telemetry_api_v1_telemetry_get",
    "parameters": [
     {
      "in": "query",
      "name": "device_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Device Id"
      }
     },
     {
      "in": "query",
      "name": "series_prefix",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Series Prefix"
      }
     },
     {
      "in": "query",
      "name": "since_ns",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Since Ns"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 3600,
       "maximum": 20000,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/TelemetrySample"
         },
         "title": "Response Telemetry Api V1 Telemetry Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Telemetry"
   }
  },
...
  "/api/v1/trace/available": {
   "get": {
    "operationId": "trace_available_api_v1_trace_available_get",
    "parameters": [
     {
      "in": "query",
      "name": "groups",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Groups"
      }
     },
     {
      "in": "query",
      "name": "with_format",
      "required": false,
      "schema": {
       "default": false,
       "title": "With Format",
       "type": "boolean"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Trace Available Api V1 Trace Available Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Trace Available"
   }
  },
...
  "/api/v1/trace/start": {
   "post": {
    "operationId": "trace_start_api_v1_trace_start_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TraceStart"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Trace Start Api V1 Trace Start Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Trace Start"
   }
  },
...
  "/api/v1/usbmon/start": {
   "post": {
    "operationId": "usbmon_start_api_v1_usbmon_start_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UsbmonStart"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Usbmon Start Api V1 Usbmon Start Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Usbmon Start"
   }
  },
...
 }
}